requirements.txt
# Core dependencies
pandas>=2.0.0
datasets>=2.14.0
tqdm>=4.65.0
python-dotenv>=1.0.1
numpy==1.26.4
scikit-learn==1.3.2

# Prompt injection detection libraries
llm-guard>=0.3.0

# Type hints support
typing-extensions>=4.5.0

# Data handling
dataclasses>=0.6